backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
numpy>=2.1.0,<3.0.0
pillow>=11.0.0,<12.0.0
opencv-python>=4.10.0,<5.0.0
websockets>=13.0,<15.0